+++ src/main/java/net/fiv/data_base/entities/LoginTable.java
@@ -33,8 +33,8 @@ public class LoginTable implements Table{
     @DatabaseField(dataType = DataType.STRING)
     private String offHand;
 
-    @DatabaseField(dataType = DataType.FLOAT)
-    private float xp;
+    @DatabaseField(dataType = DataType.INTEGER)
+    private int xp;
 
     public LoginTable(){
     }
@@ -79,11 +79,11 @@ public class LoginTable implements Table{
         this.inventory = inventory;
     }
 
-    public float getXp() {
+    public int getXp() {
         return xp;
     }
 
-    public void setXp(float xp) {
+    public void setXp(int xp) {
         this.xp = xp;
     }
