+++ src/main/java/net/fiv/data_base/entities/LogoutTable.java
@@ -33,8 +33,8 @@ public class LogoutTable implements Table{
     @DatabaseField(dataType = DataType.STRING)
     private String offHand;
 
-    @DatabaseField(dataType = DataType.FLOAT)
-    private float xp;
+    @DatabaseField(dataType = DataType.INTEGER)
+    private int xp;
 
     public LogoutTable(){
     }
@@ -79,7 +79,7 @@ public class LogoutTable implements Table{
         this.inventory = inventory;
     }
 
-    public float getXp() {
+    public int getXp() {
         return xp;
     }
 
@@ -91,7 +91,7 @@ public class LogoutTable implements Table{
         this.armor = armor;
     }
 
-    public void setXp(float xp) {
+    public void setXp(int xp) {
         this.xp = xp;
     }
